ch03 程序正常运行

+++ ch03-praseclass/classfile/attr_line_number_table.go
@@ -1,21 +1,21 @@
 package classfile
 
-type LineNumberTableAttribute struct{
-	lineNumberTable []*LineNumberTableEntry	
+type LineNumberTableAttribute struct {
+	lineNumberTable []*LineNumberTableEntry
 }
 
-type LineNumberTableEntry struct{
-	startPc uint16
+type LineNumberTableEntry struct {
+	startPc    uint16
 	lineNumber uint16
 }
 
-func (self *LineNumberTableAttribute) readInfo(reader *ClassReader){
-	lineNumberTableLength := reader.readUint32()
+func (self *LineNumberTableAttribute) readInfo(reader *ClassReader) {
+	lineNumberTableLength := reader.readUint16()
 	self.lineNumberTable = make([]*LineNumberTableEntry, lineNumberTableLength)
-	for i := range self.lineNumberTable{
+	for i := range self.lineNumberTable {
 		self.lineNumberTable[i] = &LineNumberTableEntry{
-			startPc : reader.readUint16(),
-			lineNumber : reader.readUint16(),
+			startPc:    reader.readUint16(),
+			lineNumber: reader.readUint16(),
 		}
 	}
 }
